+++ out/taskProvider.js
@@ -15,9 +15,11 @@ const vscode = require("vscode");
 const utils_1 = require("./utils");
 const EXTENSION_NAME = "C_Cpp_Runner";
 class TaskProvider {
-    constructor(settingsProvider, propertiesProvider) {
+    constructor(settingsProvider, propertiesProvider, buildMode, architectureMode) {
         this.settingsProvider = settingsProvider;
         this.propertiesProvider = propertiesProvider;
+        this.buildMode = buildMode;
+        this.architectureMode = architectureMode;
         const extDirectory = path.dirname(__dirname);
         const templateDirectory = path.join(extDirectory, "src", "templates");
         this.tasksFile = path.join(templateDirectory, "tasks_template.json");
@@ -32,26 +34,27 @@ class TaskProvider {
     provideTasks() {
         return this.getTasks();
     }
-    getTasks(ignoreLanguage = false) {
+    getTasks() {
         const editor = vscode.window.activeTextEditor;
-        let language;
-        if (false === ignoreLanguage) {
-            language = utils_1.getLanguageFromEditor(editor, this.propertiesProvider.workspacePath);
+        const language = utils_1.getLanguageFromEditor(editor, this.propertiesProvider.workspacePath);
+        this.setTasksDefinition(language);
+        if (this.tasks === undefined) {
+            return [];
         }
-        else {
-            language = utils_1.Languages.c;
-        }
-        let configJson = utils_1.readJsonFile(this.tasksFile);
+        return this.tasks;
+    }
+    setTasksDefinition(language) {
+        const configJson = utils_1.readJsonFile(this.tasksFile);
         if (undefined === configJson) {
             return [];
         }
         this.tasks = [];
-        for (let taskJson of configJson.tasks) {
-            if ("shell" !== taskJson.type) {
+        for (const taskJson of configJson.tasks) {
+            if (taskJson.type !== "shell") {
                 continue;
             }
             if (undefined !== taskJson.options) {
-                if (true === taskJson.options.hide) {
+                if (taskJson.options.hide === true) {
                     continue;
                 }
             }
@@ -69,25 +72,38 @@ class TaskProvider {
         return this.tasks;
     }
     updateTaskBasedOnSettings(taskJson, language) {
+        const settings = this.settingsProvider;
         taskJson.args[1] = `--file=${this.makefileFile}`;
-        taskJson.args.push(`ENABLE_WARNINGS=${+this.settingsProvider.enableWarnings}`);
-        taskJson.args.push(`WARNINGS="${this.settingsProvider.warnings}"`);
-        taskJson.args.push(`WARNINGS_AS_ERRORS=${+this.settingsProvider.warningsAsError}`);
-        taskJson.args.push(`C_COMPILER=${this.settingsProvider.compilerPathC}`);
-        taskJson.args.push(`CPP_COMPILER=${this.settingsProvider.compilerPathCpp}`);
+        taskJson.args.push(`COMPILATION_MODE=${this.buildMode}`);
+        taskJson.args.push(`EXECUTABLE_NAME=out${this.buildMode}`);
         taskJson.args.push(`LANGUAGE_MODE=${language}`);
-        taskJson.args.push(`C_STANDARD=${this.settingsProvider.standardC}`);
-        taskJson.args.push(`CPP_STANDARD=${this.settingsProvider.standardCpp}`);
-        if (this.settingsProvider.compilerArgs) {
-            taskJson.args.push(`COMPILER_ARGS=${this.settingsProvider.compilerArgs}`);
+        const includesClean = taskJson.label.includes(utils_1.Tasks.clean);
+        const includesRun = taskJson.label.includes(utils_1.Tasks.run);
+        if (!includesClean && !includesRun) {
+            taskJson.args.push(`ENABLE_WARNINGS=${+settings.enableWarnings}`);
+            taskJson.args.push(`WARNINGS="${settings.warnings}"`);
+            taskJson.args.push(`WARNINGS_AS_ERRORS=${+settings.warningsAsError}`);
+            if (language === utils_1.Languages.c) {
+                taskJson.args.push(`C_COMPILER=${settings.compilerPathC}`);
+                taskJson.args.push(`C_STANDARD=${settings.standardC}`);
+            }
+            else {
+                taskJson.args.push(`CPP_COMPILER=${settings.compilerPathCpp}`);
+                taskJson.args.push(`CPP_STANDARD=${settings.standardCpp}`);
+            }
+            if (settings.compilerArgs) {
+                taskJson.args.push(`COMPILER_ARGS=${settings.compilerArgs}`);
+            }
+            if (settings.linkerArgs) {
+                taskJson.args.push(`LINKER_ARGS=${settings.linkerArgs}`);
+            }
+            if (settings.includePaths) {
+                taskJson.args.push(`INCLUDE_PATHS=${settings.includePaths}`);
+            }
+            const architectureStr = this.architectureMode === utils_1.Architectures.x64 ? "64" : "32";
+            taskJson.args.push(`ARCHITECTURE=${architectureStr}`);
         }
-        if (this.settingsProvider.linkerArgs) {
-            taskJson.args.push(`LINKER_ARGS=${this.settingsProvider.linkerArgs}`);
-        }
-        if (this.settingsProvider.includePaths) {
-            taskJson.args.push(`INCLUDE_PATHS=${this.settingsProvider.includePaths}`);
-        }
-        taskJson.command = this.settingsProvider.makePath;
+        taskJson.command = settings.makePath;
     }
 }
 exports.TaskProvider = TaskProvider;